+++ sdk/core/core-client-rest/src/common.ts
@@ -323,27 +323,9 @@ export interface AdditionalPolicyConfig {
 }
 
 /**
- * The common set of options that high-level clients are expected to expose.
- */
-export interface CommonClientOptions extends PipelineOptions {
-  /**
-   * Specify a custom HttpClient when making requests.
-   */
-  httpClient?: HttpClient;
-  /**
-   * Option to allow calling HTTP (insecure) endpoints.
-   */
-  allowInsecureConnection?: boolean;
-  /**
-   * Additional policies to include in the HTTP pipeline.
-   */
-  additionalPolicies?: AdditionalPolicyConfig[];
-}
-
-/**
  * General options that a Rest Level Client can take
  */
-export type ClientOptions = CommonClientOptions & {
+export type ClientOptions = PipelineOptions & {
   /**
    * Credentials information
    */
@@ -370,6 +352,18 @@ export type ClientOptions = CommonClientOptions & {
    * Options for setting a custom apiVersion.
    */
   apiVersion?: string;
+  /**
+   * Option to allow calling http (insecure) endpoints
+   */
+  allowInsecureConnection?: boolean;
+  /**
+   * Additional policies to include in the HTTP pipeline.
+   */
+  additionalPolicies?: AdditionalPolicyConfig[];
+  /**
+   * Specify a custom HttpClient when making requests.
+   */
+  httpClient?: HttpClient;
   /**
    * Options to configure request/response logging.
    */