+++ sdk/schemaregistry/schema-registry/src/operations.ts
@@ -12,7 +12,11 @@ import type {
 } from "./models.js";
 import { buildContentType, convertSchemaIdResponse, convertSchemaResponse } from "./conversions.js";
 import type { SchemaRegistryClient } from "./clientDefinitions.js";
-import { createRestError, operationOptionsToRequestParameters } from "@azure-rest/core-client";
+import {
+  createRestError,
+  operationOptionsToRequestParameters,
+  type OperationOptions as RestOperationOptions,
+} from "@azure-rest/core-client";
 
 export async function registerSchema(
   context: SchemaRegistryClient,
@@ -20,7 +24,7 @@ export async function registerSchema(
   options: RegisterSchemaOptions = {},
 ): Promise<SchemaProperties> {
   const { groupName, name: schemaName, definition: schemaContent, format } = schema;
-  const requestParameters = operationOptionsToRequestParameters(options);
+  const requestParameters = operationOptionsToRequestParameters(options as RestOperationOptions);
   const response = await context
     .path("/$schemaGroups/{groupName}/schemas/{schemaName}", groupName, schemaName)
     .put({
@@ -45,7 +49,7 @@ export async function getSchemaProperties(
   options: GetSchemaPropertiesOptions = {},
 ): Promise<SchemaProperties> {
   const { groupName, name: schemaName, definition: schemaContent, format } = schema;
-  const requestParameters = operationOptionsToRequestParameters(options);
+  const requestParameters = operationOptionsToRequestParameters(options as RestOperationOptions);
   const response = await context
     .path("/$schemaGroups/{groupName}/schemas/{schemaName}:get-id", groupName, schemaName)
     .post({
@@ -67,7 +71,7 @@ export async function getSchemaById(
 ): Promise<Schema> {
   const response = await context
     .path("/$schemaGroups/$schemas/{id}", schemaId)
-    .get(operationOptionsToRequestParameters(options ?? {}));
+    .get(operationOptionsToRequestParameters((options ?? {}) as RestOperationOptions));
 
   if (isUnexpected(response)) {
     throw createRestError(response);
@@ -90,7 +94,7 @@ export async function getSchemaByVersion(
       name,
       version,
     )
-    .get(operationOptionsToRequestParameters(options ?? {}));
+    .get(operationOptionsToRequestParameters((options ?? {}) as RestOperationOptions));
 
   if (isUnexpected(response)) {
     throw createRestError(response);