+++ sdk/tables/data-tables/src/models.ts
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import type { CommonClientOptions } from "@azure-rest/core-client";
+import type { CommonClientOptions, OperationOptions } from "@azure/core-client";
 import type { TableGetAccessPolicyHeaders, TableInsertEntityHeaders } from "./generatedModels.js";
-import type { OperationOptions } from "./operationOptions.js";
 
 /**
  * Represents the Create or Delete Entity operation to be included in a Transaction request