+++ sdk/core/core-client-rest/CHANGELOG.md
@@ -5,11 +5,12 @@
 ### Features Added
 
 - Added an optional `internal` property to `ClientOptions` for `getClient`, carrying a `noDefaultAcceptHeader` flag (defaults to `false`). When set to `true`, the client no longer adds a default `Accept: application/json` request header to operations that do not otherwise specify one, allowing operations that expect no response body to omit the `Accept` header. This option is intended for use by generated clients. [PR #39291](https://github.com/Azure/azure-sdk-for-js/pull/39291)
-- Added `CommonClientOptions` and compatibility handling for package-local `customHeaders` and `credentialScopes` aliases used by TypeSpec-migrated convenience clients. [#39593](https://github.com/Azure/azure-sdk-for-js/issues/39593)
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+- Added support for normalizing package-local `customHeaders` aliases. [#39593](https://github.com/Azure/azure-sdk-for-js/issues/39593)
 
 ### Other Changes